+++ mitra_models/migrations/V0081__activitypub_object__add_profile_id.sql
@@ -1,0 +1,1 @@
+ALTER TABLE activitypub_object ADD COLUMN profile_id UUID UNIQUE REFERENCES actor_profile (id) ON DELETE CASCADE;