+++ mitra_models/migrations/schema.sql
@@ -277,5 +277,6 @@ CREATE TABLE subscription (
 CREATE TABLE activitypub_object (
     object_id VARCHAR(2000) PRIMARY KEY,
     object_data JSONB NOT NULL,
+    profile_id UUID UNIQUE REFERENCES actor_profile (id) ON DELETE CASCADE,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );